// frontend/src/models/Request.ts
export type Customer = {
  _id: string;
  firstname: string;
  lastname: string;
  email: string;
  password: string;
  phone: string;
  streetAddress: string;
  postalCode: string;
  city: string;
  country: string;
};

export type Seller = {
  _id: string;
  firstname: string;
  lastname: string;
};


export type Request = {
  _id: string;
  customerId: Customer;

  dimensions: {
    height: number;
    width: number;
    length: number;
    bedLength?: number;
    bedHeight?: number;

    doorDimensions: {
      doorPlacement: "right" | "left" | "aft";
      doorWidth: number;
      doorHeight: number;
    };

    windowOptions: {
      rightWindowHeight?: number;
      rightWindowLength?: number;
      leftWindowHeight?: number;
      leftWindowLength?: number;
      bedWindowHeight?: number;
      bedWindowLength?: number;
      rearWindowHeight?: number;
      rearWindowLength?: number;
    };
  };

  preferences: string;

  agreedPrice: number;

  status: "requested" | "assigned" | "approved" | "paid";

  sellerId: Seller;
  createdAt: string;
  updatedAt: string;
};

export interface IRequestCreate {
  customerId: string;

  dimensions: {
    height: number;
    width: number;
    length: number;
    bedLength?: number;
    bedHeight?: number;

    doorDimensions: {
      doorPlacement: "right" | "left" | "aft";
      doorWidth: number;
      doorHeight: number;
    };

    windowOptions: {
      rightWindowHeight?: number;
      rightWindowLength?: number;
      leftWindowHeight?: number;
      leftWindowLength?: number;
      bedWindowHeight?: number;
      bedWindowLength?: number;
      rearWindowHeight?: number;
      rearWindowLength?: number;
    };
  };

  preferences: string;

  agreedPrice?: number;

  status: "requested" | "assigned" | "approved" | "paid";
}

export interface IRequestUpdate {
  _id: string;
  customerId: string;

  dimensions: {
    height: number;
    width: number;
    length: number;
    bedLength?: number;
    bedHeight?: number;

    doorDimensions: {
      doorPlacement: "right" | "left" | "aft";
      doorWidth: number;
      doorHeight: number;
    };

    windowOptions: {
      rightWindowHeight?: number;
      rightWindowLength?: number;
      leftWindowHeight?: number;
      leftWindowLength?: number;
      bedWindowHeight?: number;
      bedWindowLength?: number;
      rearWindowHeight?: number;
      rearWindowLength?: number;
    };
  };

  preferences: string;

  status: "requested" | "assigned" | "approved" | "paid";

  sellerId: string;
}

export interface IRequestUpdateSeller{
  sellerId: string;
}

export interface IRequestUpdateStatus{
    status: "requested" | "assigned"| "approved" | "paid";
}

export interface IRequestUpdatePrice{
    agreedPrice: number;
}